pyproject.toml
[project]
name = "easy-ai-summary-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.1"
dependencies = [
    "black>=25.1.0",
    "ipykernel>=6.30.1",
    "isort>=6.0.1",
    "langchain-tavily>=0.2.11",
    "langchain[openai]>=0.3.27",
    "langgraph>=0.6.6",
    "langgraph-cli[inmem]>=0.4.0",
    "langsmith>=0.4.21",
    "python-dotenv>=1.1.1",
    "typer>=0.17.3",
]
